fix link & crud

+++ src/components/modules/StoreContent/StoreProduct.jsx
@@ -2,48 +2,86 @@ import React, { useEffect, useState } from "react";
 import style from "./storeContent.module.css";
 import { TableItem } from "../../base/TableItem/TableItem";
 import axios from "axios";
-import { useHistory } from "react-router";
+import arrowOrder from "../../../assets/icon/arrowOrder.svg";
+import { LoaderPage } from "../../base/LoaderPage/LoaderPage";
+import swal from "sweetalert";
 
 export const StoreProduct = () => {
-  const history = useHistory();
+  const [Loading, setLoading] = useState(false);
   const [data, setData] = useState([]);
-  const [keyword, setKeyword] = useState()
-  const link = "http://localhost:4000/product/";
+  const [keyword, setKeyword] = useState("");
+  const [order, setOrder] = useState("");
+  const [limit, setlimit] = useState(5);
+  const [orderBy, setorderBy] = useState("id_product");
+  const [page, setPage] = useState("");
+  const [totalData, setTotalData] = useState("");
+  let totalPage = "";
 
-  const handleSearch = (e) => {
-    setKeyword(e.target.value)
+  const handleOrder = (e) => {
+    setOrder(e.target.value);
+  };
+
+  const handleOrderBy = (ord) => {
+    setorderBy(ord);
+  };
+
+  let link = `http://localhost:4000/product/?order=${order}&orderBy=${orderBy}&page=${page}`;
+
+  if (limit != "") {
+    link += `&limit=${limit}`;
   }
 
-  const handleDelete = (e) => {
-    const id_product = e.target.value;
-    axios.delete(`http://localhost:4000/product/${id_product}`)
-    history.push('/product')
+  if (keyword != "") {
+    link += `&keyword=${keyword}`;
+  }
+
+  totalPage = Math.floor(totalData / limit);
+  if(totalData%limit > 0){
+    totalPage +=1;
+  }
+
+  const iteration = [];
+  for (let i = 1; i <= totalPage; i++) {
+    iteration.push(i);
   }
 
   useEffect(() => {
     getAllData();
-  }, []);
+  }, [keyword, order, orderBy, limit, page]);
 
   const getAllData = () => {
+    console.log(link);
     axios
       .get(link)
       .then((response) => {
-        // setData(response.data)
         setData(response.data.data);
       })
       .catch((err) => {
         console.log(err);
       });
+
+    axios.get("http://localhost:4000/product").then((response) => {
+      setTotalData(response.data.amount)
+    });
   };
 
-  // if(keyword){
-  //   getAllData( link += `?keyword=${keyword}`)
-  // }else{
-  //   getAllData(link)
-  // }
+  const handleDelete = (e) => {
+    const id_product = e;
+    swal("Berhasil", "Product sukses ditambahkan", "success").then((value) => {
+      if (value | (value == false)) {
+        setLoading(true);
+        axios.delete(`http://localhost:4000/product/${id_product}`);
+        setTimeout(() => {
+          getAllData();
+          setLoading(false);
+        }, 500);
+      }
+    });
+  };
 
   return (
     <div className={`container pt-3 pb-5 ${style.content}`}>
+      {Loading && <LoaderPage />}
       <h2>My Products</h2>
       <div className="d-flex table-order fw-bold flex-wrap">
         <input
@@ -51,7 +89,7 @@ export const StoreProduct = () => {
           class="table-order-radio-all d-none"
           name="table-order-radio"
           id="table-order-all"
-          // checked
+          defaultChecked
         />
         <input
           type="radio"
@@ -86,39 +124,117 @@ export const StoreProduct = () => {
             class="form-control"
             placeholder="Search"
             aria-describedby="addon-wrapping"
-            onChange={e => handleSearch(e)}
+            onChange={(e) => setKeyword(e.target.value)}
           ></input>
         </div>
       </div>
       <div className="table-order-product mt-5">
+        <div className="d-flex">
+          {/* Order */}
+          <div className="w-25 ms-2 mt-2 mb-2">
+            <span>Order : </span>
+            <select
+              class="form-select"
+              aria-label="Default select example"
+              onChange={(e) => handleOrder(e)}
+            >
+              <option value="ASC" selected>
+                Ascending
+              </option>
+              <option value="DESC">Descending</option>
+            </select>
+          </div>
+          {/* Amount of data */}
+          <div className="w-25 ms-2 mt-2 mb-2">
+            <span>Amount of data : </span>
+            <select
+              class="form-select"
+              aria-label="Default select example"
+              onChange={(e) => setlimit(e.target.value)}
+            >
+              <option value="5">5 data / page</option>
+              <option value="10">10 data / page</option>
+              <option value="15">15 data / page</option>
+            </select>
+          </div>
+          {/* Page */}
+          <div className="w-25 ms-2 mt-2 mb-2">
+            <span>Page 1 of {totalPage} page</span>
+            <select
+              class="form-select"
+              aria-label="Default select example"
+              onChange={(e) => setPage(e.target.value)}
+            >
+              {iteration.map((numberPage) => (
+                <option value={numberPage}>{numberPage}</option>
+              ))}
+            </select>
+          </div>
+        </div>
+        {/* <Pagination total="1"/>, */}
         <table style={{ width: "100%" }}>
           <thead>
             <tr>
-              <th style={{ width: "40%" }}>
+              <th style={{ width: "40%", position: "relative" }}>
                 Product name
-                <img src="" alt="" />
+                <button
+                  className="btn"
+                  style={{
+                    "box-shadow": "none",
+                    position: "absolute",
+                    right: 0,
+                    bottom: "20%",
+                  }}
+                  onClick={() => handleOrderBy("product_name")}
+                >
+                  <img src={arrowOrder} alt="" />
+                </button>
               </th>
-              <th style={{ width: "20%" }}>
+              <th style={{ width: "20%", position: "relative" }}>
                 Price
-                <img src="" alt="" />
+                <button
+                  className="btn"
+                  style={{
+                    "box-shadow": "none",
+                    position: "absolute",
+                    right: 0,
+                    bottom: "20%",
+                  }}
+                  onClick={() => handleOrderBy("price")}
+                >
+                  <img src={arrowOrder} alt="" />
+                </button>
               </th>
-              <th style={{ width: "20%" }}>
+              <th style={{ width: "20%", position: "relative" }}>
                 Stock
-                <img src="" alt="" />
+                <button
+                  className="btn"
+                  style={{
+                    "box-shadow": "none",
+                    position: "absolute",
+                    right: 0,
+                    bottom: "20%",
+                  }}
+                  onClick={() => handleOrderBy("quantity")}
+                >
+                  <img src={arrowOrder} alt="" />
+                </button>
               </th>
               <th style={{ width: "20%" }}>Action</th>
             </tr>
           </thead>
           <tbody>
-            {data.map((item) => (
-              <TableItem
-                title={item.product_name}
-                price={item.price}
-                quantity={item.quantity}
-                idProduct={item.id_product}
-                click={(e) => handleDelete(e)}
-              />
-            ))}
+            {data &&
+              data.map((item) => (
+                <TableItem
+                  key={item.id_product}
+                  title={item.product_name}
+                  price={item.price}
+                  quantity={item.quantity}
+                  idProduct={item.id_product}
+                  click={(e) => handleDelete(e)}
+                />
+              ))}
           </tbody>
         </table>
       </div>
